refactor: move main components to views and remove unused ones
