--- .github/FUNDING.yml
@@ -1,2 +1,0 @@
-
-custom: ['https://www.buymeacoffee.com/UJGejcNli']
