+++ src/modules/overlay/Selector.vue
@@ -99,7 +99,7 @@ export default {
   width: 100%;
   height: 100%;
   pointer-events: none;
-  z-index: 2147483647;
+  z-index: 2147483646;
 }
 
 .selector {
