+++ src/views/Recording.vue
